refactor(black-box): move the ALM delivery error prefixes into the messaging folder
The prefixes belong to the messaging feature that throws with them, and keeping
them in the runtime's top directory pushed it past the layout density and
feature-prefix review thresholds. The filename now matches its primary export.
